+++ src/handler/handler.ts
@@ -2,7 +2,7 @@ import { Client, middleware } from '@line/bot-sdk'
 import express from 'express'
 import { loadConfig } from '../config/config'
 import { managerLineHandler } from './manager_handler/manager_handler'
-import { recipientLineHandler } from './recipient_handler/recipient_line'
+import { recipientLineHandler } from './recipient_handler/recipient_handler'
 import { newFirestore } from '../lib/firestore/firestore'
 import { newStorage } from '../lib/storage/storage'
 import admin from 'firebase-admin'
@@ -31,11 +31,11 @@ newFirestore()
 newStorage()
 
 app.post('/manager-line', managerMiddleware, (req, res) =>
-  new managerLineHandler(managerClient).handle(req, res),
+  new managerLineHandler(managerClient, recipientClient).handle(req, res),
 ) //* without [(req, res) =>] it was not working. temporary fix.
 
 app.post('/recipient-line', recipientMiddleware, (req, res) =>
-  new recipientLineHandler(recipientClient, managerClient).handle(req, res),
+  new recipientLineHandler(managerClient, recipientClient).handle(req, res),
 )
 
 // TODO: 仕様が固まり次第着手します
