+++ src/handler/recipient_handler/post_handler.test.ts
@@ -6,6 +6,7 @@ import { TextTemplate } from '../../lib/line/template'
 import { Post } from '../../types/post'
 import { Recipient } from '../../types/recipient'
 import {
+  PostPreview,
   askBody,
   askImage,
   askSubjectAgain,
@@ -15,7 +16,7 @@ import {
   confirmSubject,
   discardPost,
 } from './post'
-import { reactPostText } from './post_line'
+import { reactPostText } from './post_handler'
 import admin from 'firebase-admin'
 
 const getPost = (status: postStatusType): Post => {
@@ -51,64 +52,88 @@ jest.mock('../../lib/firestore/post', () => ({
   deletePost: jest.fn(),
 }))
 
+jest.mock('../../lib/firestore/manager', () => ({
+  getManagers: jest.fn(() => []),
+}))
+
+jest.mock('../../lib/line/line', () => ({
+  Push: jest.fn(),
+}))
+
 jest.mock('../../lib/firestore/recipient', () => ({
   updateRecipient: jest.fn(),
 }))
 
-describe('recipient_line/post_line 記事投稿', () => {
+describe('recipient_handler/post_handler 記事投稿', () => {
   admin.initializeApp()
   newFirestore()
+  const managerClient = undefined as any
   const recipient = getRecipient()
+
   it(':主題入力', async () => {
     const post = getPost(postStatus.INPUT_SUBJECT)
-    expect(await reactPostText('主題', recipient, post)).toMatchObject([confirmSubject('主題')])
+    expect(await reactPostText(managerClient, '主題', recipient, post)).toMatchObject([
+      confirmSubject('主題'),
+    ])
   })
 
   describe(':主題確認', () => {
     it(':YES', async () => {
       const post = getPost(postStatus.CONFIRM_SUBJECT)
-      expect(await reactPostText(keyword.YES, recipient, post)).toMatchObject([askBody()])
+      expect(await reactPostText(managerClient, keyword.YES, recipient, post)).toMatchObject([
+        askBody(),
+      ])
     })
     it(':NO', async () => {
       const post = getPost(postStatus.CONFIRM_SUBJECT)
-      expect(await reactPostText(keyword.NO, recipient, post)).toMatchObject([askSubjectAgain()])
+      expect(await reactPostText(managerClient, keyword.NO, recipient, post)).toMatchObject([
+        askSubjectAgain(),
+      ])
     })
     it(':その他', async () => {
       const post = getPost(postStatus.CONFIRM_SUBJECT)
-      expect(await reactPostText('hoge', recipient, post)).toMatchObject([
+      expect(await reactPostText(managerClient, 'hoge', recipient, post)).toMatchObject([
         TextTemplate(phrase.yesOrNo),
       ])
     })
   })
   it(':本文入力', async () => {
     const post = getPost(postStatus.INPUT_BODY)
-    expect(await reactPostText('本文', recipient, post)).toMatchObject([confirmBody('本文')])
+    expect(await reactPostText(managerClient, '本文', recipient, post)).toMatchObject([
+      confirmBody('本文'),
+    ])
   })
   it(':本文確認', async () => {
     const post = getPost(postStatus.CONFIRM_BODY)
-    expect(await reactPostText(keyword.YES, recipient, post)).toMatchObject([askImage()])
+    expect(await reactPostText(managerClient, keyword.YES, recipient, post)).toMatchObject([
+      askImage(),
+    ])
   })
 
   //* 画像入力はテストが難しい。
   it(':画像完了', async () => {
     const post = getPost(postStatus.INPUT_IMAGE)
-    expect(await reactPostText(keyword.FINISH_IMAGE, recipient, post)).toMatchObject([
-      confirmPost(),
-    ])
+    expect(await reactPostText(managerClient, keyword.FINISH_IMAGE, recipient, post)).toMatchObject(
+      [PostPreview(post.subject, post.body, post.images), confirmPost()],
+    )
   })
 
   describe(':投稿確認', () => {
     it(':決定', async () => {
       const post = getPost(postStatus.CONFIRM_SUBMIT)
-      expect(await reactPostText(keyword.DECIDE, recipient, post)).toMatchObject([completePost()])
+      expect(await reactPostText(managerClient, keyword.DECIDE, recipient, post)).toMatchObject([
+        completePost(),
+      ])
     })
     it(':破棄', async () => {
       const post = getPost(postStatus.CONFIRM_SUBMIT)
-      expect(await reactPostText(keyword.DISCARD, recipient, post)).toMatchObject([discardPost()])
+      expect(await reactPostText(managerClient, keyword.DISCARD, recipient, post)).toMatchObject([
+        discardPost(),
+      ])
     })
     it(':その他', async () => {
       const post = getPost(postStatus.CONFIRM_SUBMIT)
-      expect(await reactPostText('hoge', recipient, post)).toMatchObject([
+      expect(await reactPostText(managerClient, 'hoge', recipient, post)).toMatchObject([
         TextTemplate(phrase.aOrb(keyword.DECIDE, keyword.DISCARD)),
       ])
     })
