+++ src/handler/recipient_handler/recipient_handler.test.ts
@@ -1,7 +1,7 @@
 import { WebhookEvent } from '@line/bot-sdk'
 import { recipientStatus, recipientStatusType } from '../../consts/constants'
 import { Recipient } from '../../types/recipient'
-import { handleEvent } from './recipient_line'
+import { handleEvent } from './recipient_handler'
 
 const getRecipient = (
   recipientGroupId: string,
@@ -35,11 +35,12 @@ jest.mock('../../lib/firestore/recipient', () => ({
 
 describe('recipient_line/recipient_line フォロー', () => {
   const event = getEvent('follow')
-  const client = undefined as any
+  const managerClient = undefined as any
+  const recipientClient = undefined as any
   it(':名前入力から再開', async () => {
     getRecipient_.mockImplementation(() => getRecipient('', '', recipientStatus.NONE))
 
-    expect(await handleEvent(client, event)).toMatchObject([
+    expect(await handleEvent(managerClient, recipientClient, event)).toMatchObject([
       {
         type: 'text',
         text: '友だち追加ありがとうございます。\nこのアカウントでは、文章や画像をチャット送っていただくだけで記事投稿が出来ます。',
@@ -51,7 +52,7 @@ describe('recipient_line/recipient_line フォロー', () => {
   it(':団体ID入力から再開', async () => {
     getRecipient_.mockReturnValue(getRecipient('', '太郎', recipientStatus.NONE))
 
-    expect(await handleEvent(client, event)).toMatchObject([
+    expect(await handleEvent(managerClient, recipientClient, event)).toMatchObject([
       {
         type: 'text',
         text: '友だち追加ありがとうございます。\nこのアカウントでは、文章や画像をチャット送っていただくだけで記事投稿が出来ます。',
@@ -63,7 +64,7 @@ describe('recipient_line/recipient_line フォロー', () => {
   it(':即復帰', async () => {
     getRecipient_.mockReturnValue(getRecipient('rg-0001', '太郎', recipientStatus.NONE))
 
-    expect(await handleEvent(client, event)).toMatchObject([
+    expect(await handleEvent(managerClient, recipientClient, event)).toMatchObject([
       {
         text: '「太郎」さん、おかえりなさい。',
         type: 'text',
